long to int

+++ src/main/java/app/controller/RegisterNewCALController.java
@@ -39,7 +39,7 @@ public class RegisterNewCALController {
      * @param answer
      * @return boolean
      */
-    public boolean registerNewCAL(String labId, String labName, long phoneNumber, String address, int tin, boolean answer){
+    public boolean registerNewCAL(String labId, String labName, long phoneNumber, String address, long tin, boolean answer){
         this.cal = this.calStore.registerNewCAL(labId, labName, phoneNumber, address, tin, answer);
         return this.calStore.validateCAL(this.cal);
     }
